* src/preproc/html/pre-html.cc (scanArguments): Use `troff_command' instead of hardwired `"troff"'. (char_buffer::read_file): Remove unused variables. (char_buffer::write_file_troff): Ditto. (char_buffer::write_file_html): Ditto. (generateImages): Ditto. (abs): Removed. * src/utils/addftinfo/addftinfo.cc (main): Fix compilation problem.
* src/include/lib.h: Don't include `groff-getopt.h' for UWIN.

* src/libs/libgroff/font.cc (font::load_desc): Fix compiler warning.
* src/libs/libbib/index.cc (index_search_item_iterator::get_tag,
index_search_item::search1): Ditto.
* src/roff/troff/node.h (width_list): Reorder members.
* src/roff/troff/input.cc (do_define_macro): Fix compiler warning.
(lookup_warning): Ditto.
(prepend_string): Commented out.
* src/roff/troff/dictionary.cc (is_good_size): Fix compiler warning.
* src/roff/troff/reg.cc (number_value_to_ascii): Ditto.
* src/preproc/tbl/main.cc (process_format): Ditto.
* src/preproc/pic/pic.y (define_variable, reset, reset_all): Fix
compiler warnings.
* src/preproc/pic/pic.cc: Updated.
* src/preproc/eqn/lex.cc (init_table): Fix compiler warnings.
* src/preproc/eqn/over.cc (over_box::compute_metrics): Ditto.
* src/preproc/refer/command.cc (execute_command): Ditto.
* src/preproc/refer/ref.cc (sortify_title, find_month): Ditto.
* src/preproc/refer/token.cc (lookup_token): Ditto.
* src/preproc/soelim/soelim.cc: Ditto.
* src/preproc/html/pushbackbuffer.cc (ERROR): Ditto.
(isHexDigit): Commented out.
(pushBackBuffer::isString): Remove unused variable.
* src/devices/grops/ps.cc (ps_printer::special): Fix compiler
warning.
* src/devices/grops/psrm.cc (resource_manager::lookup_font,
resource_manager::read_resource_arg, read_one_of, parse_extensions):
Ditto.
* src/devices/grodvi/dvi.cc (dvi_printer::draw): Ditto.
* src/devices/grolj4/lj4.cc (lj4_font::handle_unknown_font_command,
lookup_paper_size): Ditto.
* src/devices/grohtml/post-html.cc (is_subsection): Commented out.
(text_glob, element_list): Reorder members.
* src/devices/grohtml/html-text.cc (html_text::do_pre): Remove
compiler warning.
* src/devices/grohtml/html-text.h (html_text::emit_space): Change
return value to `void'.
* src/devices/grohtml/output.cc (word_list::word_list): Fix order
of initializers.
* src/devices/grohtml/html.h (simple_output::check_newline,
simple_output::space_or_newline, simple_output::enable_newlines):
Add return value.
* src/devices/grolbp/lbp.cc (wp54charset, set_papersize): Remove
compiler warnings.
* src/utils/tfmtodit/tfmtodit.cc (tfm::load, gf::load, main): Ditto.
* src/utils/hpftodit/hpftodit.cc (File::seek, read_symbol_sets,
output_ligatures, output_charset, read_map): Ditto.
* src/utils/lookbib/lookbib.cc (main): Ditto.
* src/utils/indxbib/indxbib.cc (main, fwrite_or_die): Ditto.
(do_file): Fix typo (`=' -> `==').
* src/utils/lkbib/lkbib.cc (main): Remove compiler warning.
* src/utils/addftinfo/addftinfo.cc (main): Ditto.

+++ src/preproc/soelim/soelim.cc
@@ -30,7 +30,7 @@ Foundation, 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA. */
 #include "stringclass.h"
 #include "nonposix.h"
 
-static int include_list_length;
+static size_t include_list_length;
 static char **include_list;
 
 int compatible_flag = 0;
@@ -183,18 +183,18 @@ int do_file(const char *filename)
     {
       char *path = include_list[j];
       if (0 == strcmp(path, "."))
-      	whole_filename = filename;
+	whole_filename = filename;
       else
-        whole_filename = string(path) + "/" + filename;
+	whole_filename = string(path) + "/" + filename;
       whole_filename += '\0';
       errno = 0;
       fp = fopen(whole_filename.contents(), "r");
       if (fp != 0)
-      	break;
+	break;
       if (errno != ENOENT) {
-        error("can't open `%1': %2",
+	error("can't open `%1': %2",
 	      whole_filename.contents(), strerror(errno));
-        return 0;
+	return 0;
       }
     }
     if (j >= include_list_length)
