+++ src/utils/lkbib/lkbib.cc
@@ -126,7 +126,7 @@ int main(int argc, char **argv)
   int len;
   int count;
   for (count = 0; iter.next(&start, &len); count++) {
-    if (fwrite(start, 1, len, stdout) != len)
+    if (fwrite(start, 1, len, stdout) != (size_t)len)
       fatal("write error on stdout: %1", strerror(errno));
     // Can happen for last reference in file.
     if (start[len - 1] != '\n')
